refactor: ui changes, use more space
