+++ dash/src/components/dashboard/LeaderBoard.tsx
@@ -16,7 +16,7 @@ export default function LeaderBoard() {
 	const showTableHeader = useSettingsStore((state) => state.tableHeaders);
 
 	return (
-		<div className="flex w-fit flex-col divide-y divide-zinc-800">
+		<div className="flex w-fit flex-col gap-1">
 			{showTableHeader && <TableHeaders />}
 
 			{(!drivers || !driversTiming) &&
