+++ dash/src/components/dashboard/RadioMessage.tsx
@@ -74,7 +74,7 @@ export default function RadioMessage({ driver, capture, basePath }: Props) {
 		<motion.li
 			animate={{ opacity: 1, y: 0 }}
 			initial={{ opacity: 0, y: -20 }}
-			className={clsx("flex flex-col gap-1 p-2", { "bg-sky-800/30": favoriteDriver })}
+			className={clsx("flex flex-col gap-1 rounded-lg p-2", { "bg-sky-800/30": favoriteDriver })}
 		>
 			<time
 				className="text-sm leading-none font-medium text-gray-500"
